pyproject.toml
[tool.poetry]
name = "google-locdest"
version = "0.1.0"
description = "get location dest from google"
authors = ["amir_espahbodi <dev.amir.espahbodi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.7"
environs = "^14.1.1"
psycopg2-binary = "^2.9.10"
gunicorn = "^23.0.0"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^25.1"
ruff = "^0.10.0"
django-debug-toolbar = "^5.0.1"
uvicorn = "^0.34.0"
httpx = "^0.28.1"
whitenoise = "^6.9.0"
drf-spectacular = "^0.28.0"
faker = "^37.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
